Refactor indexer_config -> config
