+++ medusa/indexers/api.py
@@ -4,7 +4,7 @@ import os
 
 from medusa import app
 from medusa.helper.common import try_int
-from medusa.indexers.indexer_config import indexerConfig, initConfig
+from medusa.indexers.config import indexerConfig, initConfig
 from medusa.indexers.tvdbv2.fallback import PlexFallBackConfig
