+++ medusa/providers/nzb/newznab.py
@@ -22,7 +22,7 @@ from medusa.helper.common import (
 )
 from medusa.helper.encoding import ss
 from medusa.helpers.utils import split_and_strip
-from medusa.indexers.indexer_config import (
+from medusa.indexers.config import (
     INDEXER_TMDB,
     INDEXER_TVDBV2,
     INDEXER_TVMAZE,
