+++ medusa/scene_exceptions.py
@@ -14,7 +14,7 @@ import adba
 
 from medusa import app, db, helpers
 from medusa.indexers.api import indexerApi
-from medusa.indexers.indexer_config import INDEXER_TVDBV2
+from medusa.indexers.config import INDEXER_TVDBV2
 from medusa.session.core import MedusaSafeSession
 
 from six import iteritems
