+++ medusa/server/web/home/add_shows.py
@@ -13,7 +13,7 @@ from medusa.common import Quality
 from medusa.helper.common import sanitize_filename, try_int
 from medusa.helpers import get_showname_from_indexer
 from medusa.indexers.api import indexerApi
-from medusa.indexers.indexer_config import INDEXER_TVDBV2
+from medusa.indexers.config import INDEXER_TVDBV2
 from medusa.indexers.indexer_exceptions import IndexerException, IndexerUnavailable
 from medusa.server.web.core import PageTemplate
 from medusa.server.web.home.handler import Home
