+++ medusa/show/recommendations/imdb.py
@@ -9,7 +9,7 @@ import re
 from datetime import date
 from imdbpie import imdbpie
 from medusa import app, helpers
-from medusa.indexers.indexer_config import INDEXER_TVDBV2
+from medusa.indexers.config import INDEXER_TVDBV2
 from medusa.logger.adapters.style import BraceAdapter
 from medusa.session.core import MedusaSession
 from medusa.show.recommendations import ExpiringKeyValue
