+++ medusa/tv/episode.py
@@ -52,7 +52,7 @@ from medusa.helper.exceptions import (
 )
 from medusa.helper.mappings import NonEmptyDict
 from medusa.indexers.api import indexerApi
-from medusa.indexers.indexer_config import indexerConfig
+from medusa.indexers.config import indexerConfig
 from medusa.indexers.indexer_exceptions import (
     IndexerEpisodeNotFound,
     IndexerError,
