+++ views/config_notifications.mako
@@ -6,7 +6,7 @@
     from medusa.common import SKIPPED, WANTED, UNAIRED, ARCHIVED, IGNORED, SNATCHED, SNATCHED_PROPER, SNATCHED_BEST, FAILED
     from medusa.common import Quality, qualityPresets, statusStrings, qualityPresetStrings, cpu_presets, MULTI_EP_STRINGS
     from medusa.indexers.api import indexerApi
-    from medusa.indexers.indexer_config import get_trakt_indexer
+    from medusa.indexers.config import get_trakt_indexer
 %>
 <%block name="content">
 % if not header is UNDEFINED:
