+++ views/schedule.mako
@@ -3,7 +3,7 @@
     from medusa import app
     from medusa.helpers import anon_url
     from medusa.indexers.api import indexerApi
-    from medusa.indexers.indexer_config import mappings
+    from medusa.indexers.config import mappings
     from medusa import sbdatetime
     from random import choice
     import datetime
